# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RedGain membership referral platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── User dashboard ────────────────────────────────────────────────────────
...
  /payments:
    post:
      operationId: submitPayment
      tags: [payments]
      summary: Submit a payment proof
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Payment submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentRecord"
...
  /admin/payments/{id}:
    patch:
      operationId: adminReviewPayment
      tags: [admin]
      summary: Approve or reject a payment (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentReview"
      responses:
        "200":
          description: Updated payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPayment"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    RegisterInput:
      type: object
      required: [name, email, password, phone]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          description: Phone number with country code e.g. +5588992543996
        password:
          type: string
          minLength: 6
        referralCode:
          type: ["string", "null"]
        bscWallet:
          type: ["string", "null"]
          description: BSC wallet address (0x...) from which the user will send payments
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    User:
      type: object
      required: [id, name, email, role, accountStatus, referralCode, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        accountStatus:
          type: string
          enum: [pending, active, paused, lost]
        referralCode:
          type: string
        bscWallet:
          type: ["string", "null"]
          description: BSC wallet address for automatic payment detection
        membershipStartedAt:
          type: ["string", "null"]
        membershipTimerStartedAt:
          type: ["string", "null"]
        membershipExpiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
...
    PaymentInput:
      type: object
      required: [proofText, amount]
      properties:
        proofText:
          type: string
        amount:
          type: number
        proofImageUrl:
          type: ["string", "null"]
        proofImageUrls:
          type: array
          items:
            type: string
    PaymentRecord:
      type: object
      required: [id, userId, amount, paymentType, status, proofText, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
        paymentType:
          type: string
          enum: [initial, renewal]
        status:
          type: string
          enum: [pending, approved, rejected]
        proofText:
          type: string
        txHash:
          type: ["string", "null"]
          description: Blockchain transaction hash (for crypto payments)
        proofImageUrl:
          type: ["string", "null"]
        proofImageUrls:
          type: array
          items:
            type: string
        createdAt:
          type: string
        reviewedAt:
          type: ["string", "null"]
...
    AdminPayment:
      type: object
      required: [id, userId, userName, userEmail, amount, paymentType, status, proofText, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userEmail:
          type: string
        userPhone:
          type: ["string", "null"]
        amount:
          type: number
        paymentType:
          type: string
          enum: [initial, renewal]
        status:
          type: string
          enum: [pending, approved, rejected]
        proofText:
          type: string
        proofImageUrl:
          type: ["string", "null"]
        proofImageUrls:
          type: array
          items:
            type: string
        createdAt:
          type: string
        reviewedAt:
          type: ["string", "null"]
    PaymentReview:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, rejected]
        note:
          type: ["string", "null"]
...