# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * RedGain membership referral platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserAccountStatus } from './userAccountStatus';
import type { UserRole } from './userRole';

export interface User {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  role: UserRole;
  accountStatus: UserAccountStatus;
  referralCode: string;
  /**
     * BSC wallet address for automatic payment detection
     * @nullable
     */
  bscWallet?: string | null;
  /** @nullable */
  membershipStartedAt?: string | null;
  /** @nullable */
  membershipTimerStartedAt?: string | null;
  /** @nullable */
  membershipExpiresAt?: string | null;
  createdAt: string;
}